pyproject.toml
[tool.poetry]
name = "rivian-python-client"
version = "0.0.0"
description = "Rivian API Client (Unofficial)"
readme = "README.md"
authors = ["Brian Retterer <bretterer@gmail.com>"]
license = "MIT"
packages = [
    { include = "rivian", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
aiohttp = ">=3.0.0"
cryptography = ">=41.0.1,<43.0"
backports-strenum = { version = "^1.2.4", python = "<3.11" }
bleak = { version = ">=0.21,<1.0.0", optional = true }
dbus-fast = {version = "^2.11.0", optional = true, platform = "linux"}

[tool.poetry.group.dev.dependencies]
pytest = ">=7.1.2,<9.0.0"
pytest-asyncio = ">=0.18.3,<0.24.0"
python-dotenv = ">=0.20,<1.1"
aresponses = ">=2.1.5,<4.0.0"
mypy = ">=1.7.0,<2.0.0"
ruff = "^0.4"

[tool.poetry.extras]
ble = ["bleak", "dbus-fast"]

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"
pattern = "default-unprefixed"

[tool.ruff.lint.isort]
combine-as-imports = true
split-on-trailing-comma = false

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
